flask/requirements.txt
Flask
uWSGI
Werkzeug
mysql-connector
typing